添加回复功能的异常处理

+++ src/main/java/adou/community/controller/PublishController.java
@@ -1,7 +1,6 @@
 package adou.community.controller;
 
 import adou.community.dto.QuestionDTO;
-import adou.community.mapper.QuestionMapper;
 import adou.community.model.Question;
 import adou.community.model.User;
 import adou.community.service.QuestionService;
@@ -19,7 +18,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable("id") Integer id,
+    public String edit(@PathVariable("id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -39,7 +38,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam("id") Integer id,
+            @RequestParam("id") Long id,
             HttpServletRequest request,
             Model model) {
         //拿到user对象
